Changed how main() and run.sh accept filenames.
main() had assumptions that IO files were in io-files directory, and class was being executed from inside the bin directory. The reason was to simplify the command line arguments, but it was obviously not robust at all. Now main() uses pure user input as the filenames. run.sh uses directory variable to allow simpler input but be easy to change if needed.